Smarthome packages renamed to openHAB
