+++ src/main/resources/metadata/findbugs/findbugs.xml
@@ -3,9 +3,9 @@
 <!-- Custom defined detectors and bug patterns have to be included in this plugin before they are used in some of the FindBugs rule sets -->
 <FindbugsPlugin xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                 xsi:noNamespaceSchemaLocation="findbugsplugin.xsd"
-                pluginid="org.eclipse.smarthome.tools.analysis.findbugs"
+                pluginid="org.openhab.tools.analysis.findbugs"
                 defaultenabled="true"
                 provider="Eclipse SmartHome">
-  <Detector class="org.eclipse.smarthome.tools.analysis.findbugs.CustomClassNameLengthDetector" speed="fast" reports="CUST_CLASS_NAME_LENGTH" disabled="false"/>
+  <Detector class="org.openhab.tools.analysis.findbugs.CustomClassNameLengthDetector" speed="fast" reports="CUST_CLASS_NAME_LENGTH" disabled="false"/>
   <BugPattern type="CUST_CLASS_NAME_LENGTH" abbrev="CCNL" category="BAD_PRACTICE" />
 </FindbugsPlugin>
