+++ src/main/resources/rulesets/checkstyle/binding.xml
@@ -39,15 +39,15 @@
     <property name="severity" value="info"/>
   </module>
   
-  <module name="org.eclipse.smarthome.tools.analysis.checkstyle.ExportInternalPackageCheck">
+  <module name="org.openhab.tools.analysis.checkstyle.ExportInternalPackageCheck">
     <property name="severity" value="warning" />
   </module>
   
-  <module name="org.eclipse.smarthome.tools.analysis.checkstyle.MavenPomderivedInClasspathCheck">
+  <module name="org.openhab.tools.analysis.checkstyle.MavenPomderivedInClasspathCheck">
     <property name="severity" value="warning" />
   </module>
   
-  <module name="org.eclipse.smarthome.tools.analysis.checkstyle.OverridingParentPomConfigurationCheck">
+  <module name="org.openhab.tools.analysis.checkstyle.OverridingParentPomConfigurationCheck">
     <property name="severity" value="warning" />
   </module>
 
@@ -65,7 +65,7 @@
       <property name="format" value="(TODO)|(FIXME)" />
       <property name="severity" value="info" />
     </module>
-    <module name="org.eclipse.smarthome.tools.analysis.checkstyle.JavadocFilterCheck">
+    <module name="org.openhab.tools.analysis.checkstyle.JavadocFilterCheck">
       <property name="severity" value="warning" />
       <property name="scope" value="public"/>
       <property name="checkInnerUnits" value="false"/>
