+++ src/test/resources/org/openhab/tools/analysis/pmd/test/xml/AvoidMavenPomderivedInClasspath.xml
